% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumVec.R
\name{sumVec}
\alias{sumVec}
\title{Sum a vector}
\usage{
sumVec(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric(1)
}
\description{
Returns the sum of all the elements in a vector.
}
\author{
Timothee Flutre
}

