% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sumVec_cpp_ext}
\alias{sumVec_cpp_ext}
\title{Sum a vector (C/C++ external implementation)}
\usage{
sumVec_cpp_ext(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
numeric(1)
}
\description{
Return the sum of all the elements in a vector.
}
\author{
Timothee Flutre
}
