% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumVec.R
\name{sumVec_R}
\alias{sumVec_R}
\title{Sum a vector}
\usage{
sumVec_R(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
numeric(1)
}
\description{
Return the sum of all the elements in a vector.
}
\author{
Timothee Flutre
}
